Update README.AnyChart.md
Getting my feet wet with AnyData this Readme is a lifesaver and a very good manual. Great demo site as well. Thanks! To give some back, I clarified some of the text that I had to read twice or misinterpreted. 
And the link to the AnyChart building blocks seems to be broken. Can you check this? Are they still available?